+++ lib/mutant/mutator/node/send/binary.rb
@@ -21,6 +21,7 @@ module Mutant
           def dispatch
             emit(left)
             emit_left_mutations
+            emit_selector_replacement
             emit(right) unless right.type == :splat
             emit_right_mutations
           end